refactor(learnings): share sanitizeScopeGlobs so operator + distiller cap identically
Hoist sanitizeScopeGlobs (+ MAX_SCOPE_GLOBS/MAX_GLOB_LEN) from distiller into
house-rules as the single source of truth; setLearningScope now uses it, so the
operator edit path enforces the same count/length caps as the distiller —
addressing the PR critic's guard-asymmetry note. + cap regression test.
